Make consistency in output of "highlight()" for "wrap: false" (fix #111)
This patch is fix the behavior of `highlight()` with a option `wrap: false`.

If `hljs: false`, the output of `highlight()` should include a HTML class name `highlight`.
It is consistent with the case of `hljs:true`.

And also, even if `hljs: false`, the output of `highlight()` should include a PRE element and a CODE element.
It is consistent with the case of `hljs:true`, too.